requirements.txt
pandas
numpy
requests
segment-geospatial
torch
transformers
pillow